+++ ExpandWorldPrefabs/PrefabManager.cs
@@ -50,13 +50,23 @@ public class Manager
     if (info.Drops)
       SpawnDrops(zdo);
     // Original object was regenerated to apply data.
-    if (info.Remove || info.Data != "")
+    if (info.Remove || (info.Data != "" && !info.InjectData))
       DelayedRemove.Add(info.RemoveDelay, zdo, info.Remove && info.TriggerRules);
+    else if (info.InjectData)
+    {
+      var data = DataHelper.Get(info.Data);
+      if (data != null)
+      {
+        data.Write(parameters, zdo);
+        zdo.DataRevision += 100;
+        ZDOMan.instance.ForceSendZDO(zdo.m_uid);
+      }
+    }
   }
   private static void HandleSpawns(Info info, ZDO zdo, Dictionary<string, string> parameters)
   {
     // Original object must be regenerated to apply data.
-    var regenerateOriginal = !info.Remove && info.Data != "";
+    var regenerateOriginal = !info.Remove && info.Data != "" && !info.InjectData;
     if (info.Spawns.Length == 0 && info.Swaps.Length == 0 && !regenerateOriginal) return;
 
     var customData = DataHelper.Get(info.Data);